Fix imports and rename top level package
Preface imports with package name.

Rename top level package from cryptic 'src' to more descriptive
'circlegame'.
